FIX ensure arguments list is empty
If explode() fails to find a delimiter it returns an array containing the input string. In the case of a null/empty input string this lead to the arguments array containing a single null item when it should contain none, so Providers used an empty value for the first argument instead of the correct default.